LayoutTests/http/tests/security/xssAuditor/base-href-control-char-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

ALERT: This is a safe script.
